Implement single bit reading in binary_file
I changed the BufReader to an iterator because it's a lot more efficient and useful for what it's doing. Armed with that upgrade, I proceeded to implement the ability to read a single bit as a bool, and to buffer another byte, which happens automatically, but can be done manually if needed. As usual, I also added in a new test to ensure that it's working. the test just opens a sample file and ensures that it gets the first character and some of the second right.