Add comments and titles
I realized that as the code complexity grew, it would require a second to remember what each bit did. I know that if it's like that for the developer, it's so much worse for everyone else, so I added comments. For extra clarity, I added "Andromeda by Et-17" and it's purpose to the beginning of each file. Some imports got switched around as well because my new prettifier sorts the imports alphabetically.